#973 fix flickering and double click issues with Google map
