+++ src/components/views/create/Inputs.tsx
@@ -1,7 +1,7 @@
 import NameInput from './NameInput';
 import DescriptionInput from './DescriptionInput';
 import CategoryInput from './CategoryInput';
-import LocationInput from './LocationInput';
+import LocationIndex from './locationInput/Location.index';
 import ImageInput from './imageInput/ImageInput';
 import WalletAddressInput from './WalletAddressInput';
 
@@ -9,7 +9,7 @@ export {
 	NameInput,
 	DescriptionInput,
 	CategoryInput,
-	LocationInput,
+	LocationIndex,
 	ImageInput,
 	WalletAddressInput,
 };
